// src/components/WeatherInfo/type.ts
interface WeatherData {
  time: {
    local: string;
    utc: string;
  };
  weatherCode: {
    value: string;
  };
  temperature: number;
  dewPoint: number;
  feelsLike: number;
  wind: {
    direction: string;
    speed: number;
    gust: number;
  };
  relativeHumidity: number;
  pressure: {
    value: number;
    trend: number;
  };
  visibility: number;
  ceiling: number;
}

export interface WeatherInfoProps {
  data: WeatherData;
}